+++ src/main/java/com/my/foody/global/config/WebMvcConfig.java
@@ -2,13 +2,22 @@ package com.my.foody.global.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
 
 @Configuration
 @RequiredArgsConstructor
 public class WebMvcConfig implements WebMvcConfigurer {
     private final AuthInterceptor authInterceptor;
+    private final TokenSubjectArgumentResolver tokenSubjectArgumentResolver;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(tokenSubjectArgumentResolver);
+    }
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
